# examples/shared/controller/controller.go
package controller

// Controller example
type Controller struct {
}

// NewController example
func NewController() *Controller {
	return &Controller{}
}

// Message example
type Message struct {
	Message string `json:"message" example:"message"`
}

type Message2 struct {
	Message string `json:"message" example:"message"`
}